some improvements on the thumb detour

+++ ninecraft/src/main.c
@@ -64,13 +64,13 @@ void detour(void *dst_addr, void *src_addr) {
     void *protect = (void *)(((uintptr_t)dst_addr & 0xFFFFFFFE) & -page_size);
     mprotect(protect, 10, PROT_READ | PROT_WRITE | PROT_EXEC);
     uint32_t i = 0;
-    if ((((uint32_t)dst_addr & 0xFFFFFFFE) % 4) != 0) {
-        ((uint16_t *)((uint32_t)dst_addr & 0xFFFFFFFE))[i++] = 0xBF00;
+    if ((((uintptr_t)dst_addr & 0xFFFFFFFE) % 4) != 0) {
+        ((uint16_t *)((uintptr_t)dst_addr & 0xFFFFFFFE))[i++] = 0xBF00;
     }
-    ((uint16_t *)((uint32_t)dst_addr & 0xFFFFFFFE))[i++] = 0xF8DF;
-    ((uint16_t *)((uint32_t)dst_addr & 0xFFFFFFFE))[i++] = 0xF000;
-    ((uint16_t *)((uint32_t)dst_addr & 0xFFFFFFFE))[i++] = (uint32_t)src_addr & 0xFFFF;
-    ((uint16_t *)((uint32_t)dst_addr & 0xFFFFFFFE))[i++] = (uint32_t)src_addr >> 16;
+    ((uint16_t *)((uintptr_t)dst_addr & 0xFFFFFFFE))[i++] = 0xF8DF;
+    ((uint16_t *)((uintptr_t)dst_addr & 0xFFFFFFFE))[i++] = 0xF000;
+    ((uint16_t *)((uintptr_t)dst_addr & 0xFFFFFFFE))[i++] = (uint16_t)((uintptr_t)src_addr & 0xFFFF);
+    ((uint16_t *)((uintptr_t)dst_addr & 0xFFFFFFFE))[i++] = (uint16_t)((uintptr_t)src_addr >> 16);
     mprotect(protect, 10, PROT_EXEC);
 }
